Add files via upload
In this data exploration project, we leveraged SQL to extract insights from a COVID-19 dataset. We queried and aggregated data to uncover trends, such as infection rates, regional disparities, and vaccination progress. The resulting dataset was then visualized using Tableau or Power BI, creating interactive, informative dashboards and visualizations. 
This approach enabled us to provide valuable insights and a comprehensive understanding of the pandemic's impact, aiding decision-makers in informed responses and resource allocation.